clean up data space
