+++ .changeset/cool-moons-train.md
@@ -1,0 +1,7 @@
+---
+'@finos/legend-extension-dsl-data-space-studio': minor
+'@finos/legend-application-studio-bootstrap': patch
+'@finos/legend-extension-dsl-data-space': minor
+---
+
+Seperate dsl-dataspace-studio into own module.
