+++ packages/legend-application-studio-bootstrap/style/index.scss
@@ -24,6 +24,7 @@
 @import url('@finos/legend-extension-dsl-service/lib/index.css');
 @import url('@finos/legend-extension-dsl-diagram/lib/index.css');
 @import url('@finos/legend-extension-dsl-data-space/lib/index.css');
+@import url('@finos/legend-extension-dsl-data-space-studio/lib/index.css');
 @import url('@finos/legend-extension-dsl-persistence/lib/index.css');
 @import url('@finos/legend-extension-dsl-mastery/lib/index.css');
 @import url('@finos/legend-extension-dsl-text/lib/index.css');
