+++ packages/legend-application-studio-bootstrap/tsconfig.build.json
@@ -15,6 +15,8 @@
     { "path": "../legend-extension-assortment/tsconfig.build.json" },
     { "path": "../legend-extension-dsl-service/tsconfig.build.json" },
     { "path": "../legend-extension-dsl-data-space/tsconfig.build.json" },
+    { "path": "../legend-extension-dsl-data-space-studio/tsconfig.build.json" },
+
     { "path": "../legend-extension-dsl-diagram/tsconfig.build.json" },
     { "path": "../legend-extension-dsl-persistence/tsconfig.build.json" },
     { "path": "../legend-extension-dsl-mastery/tsconfig.build.json" },
