+++ packages/legend-extension-dsl-data-space-studio/.npmignore
@@ -1,0 +1,7 @@
+/build
+/style
+**/__mocks__/**
+**/__tests__/**
+/*.*
+!tsconfig.json
+!tsconfig.package.json
