+++ packages/legend-extension-dsl-data-space-studio/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @finos/legend-extension-dsl-data-space-studio
