+++ packages/legend-extension-dsl-data-space-studio/README.md
@@ -1,0 +1,3 @@
+# @finos/legend-extension-dsl-data-space-studio
+
+Legend extension for Data Space DSL Studio
