+++ packages/legend-extension-dsl-data-space-studio/jest.config.js
@@ -14,5 +14,16 @@
  * limitations under the License.
  */
 
-@forward 'data-space-viewer';
-@forward 'data-space-query-plugin';
+import { getBaseJestDOMProjectConfig } from '../../scripts/test/jest.config.base.js';
+import { loadJSON } from '@finos/legend-dev-utils/DevUtils';
+import { resolve, dirname } from 'path';
+import { fileURLToPath } from 'url';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+const packageJson = loadJSON(resolve(__dirname, './package.json'));
+
+export default getBaseJestDOMProjectConfig(
+  packageJson.name,
+  'packages/legend-extension-dsl-data-space-studio',
+);
