+++ packages/legend-extension-dsl-data-space-studio/src/__lib__/DSL_DataSpace_LegendStudioCodeSnippet.ts
