+++ packages/legend-extension-dsl-data-space-studio/src/__lib__/DSL_DataSpace_LegendStudioDocumentation.ts
@@ -14,5 +14,8 @@
  * limitations under the License.
  */
 
-@forward 'data-space-viewer';
-@forward 'data-space-query-plugin';
+export enum DSL_DATA_SPACE_LEGEND_STUDIO_DOCUMENTATION_KEY {
+  GRAMMAR_PARSER = 'dsl-dataspace.grammar.parser',
+  CONCEPT_ELEMENT_DATA_SPACE = 'dsl-dataspace.concept.element.data-space',
+  CURATED_TEMPLATE_QUERY = 'dsl-dataspace.curated-template-query',
+}
