+++ packages/legend-extension-dsl-data-space-studio/src/components/DSL_DataSpace_LegendStudioApplicationPlugin.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import packageJson from '../../../package.json' assert { type: 'json' };
+import packageJson from '../../package.json' assert { type: 'json' };
 import {
   type NewElementFromStateCreator,
   type ElementClassifier,
@@ -42,26 +42,28 @@ import {
   stub_PackageableRuntime,
   type PackageableElement,
 } from '@finos/legend-graph';
-import {
-  DataSpace,
-  DataSpaceExecutionContext,
-} from '../../graph/metamodel/pure/model/packageableElements/dataSpace/DSL_DataSpace_DataSpace.js';
-import { DSL_DATA_SPACE_LEGEND_STUDIO_DOCUMENTATION_KEY } from '../../__lib__/studio/DSL_DataSpace_LegendStudioDocumentation.js';
-import { SIMPLE_DATA_SPACE_SNIPPET } from '../../__lib__/studio/DSL_DataSpace_LegendStudioCodeSnippet.js';
 import type {
   ApplicationPageEntry,
   DocumentationEntry,
 } from '@finos/legend-application';
-import { DataSpaceIcon } from '../DSL_DataSpace_Icon.js';
-import { DataSpacePreviewState } from '../../stores/studio/DataSpacePreviewState.js';
-import {
-  DataSpacePreviewDialog,
-  DataSpacePreviewAction,
-} from './DataSpacePreviewAction.js';
 import type { PureGrammarTextSuggestion } from '@finos/legend-lego/code-editor';
-import { DataSpaceQueryAction } from './DataSpaceQueryAction.js';
+import {
+  DataSpace,
+  DataSpaceExecutionContext,
+} from '@finos/legend-extension-dsl-data-space/graph';
+import { DSL_DATA_SPACE_LEGEND_STUDIO_DOCUMENTATION_KEY } from '../__lib__/DSL_DataSpace_LegendStudioDocumentation.js';
+import { DataSpacePreviewState } from '../stores/DataSpacePreviewState.js';
+import { SIMPLE_DATA_SPACE_SNIPPET } from '../__lib__/DSL_DataSpace_LegendStudioCodeSnippet.js';
+import {
+  DATA_SPACE_STUDIO_ROUTE_PATTERN,
+  DataSpaceIcon,
+} from '@finos/legend-extension-dsl-data-space/application';
 import { DataSpaceTemplateQueryPromotionReviewer } from './DataSpaceTemplateQueryPromoteReview.js';
-import { DATA_SPACE_QUERY_ROUTE_PATTERN } from '../../__lib__/query/DSL_DataSpace_LegendQueryNavigation.js';
+import { DataSpaceQueryAction } from './DataSpaceQueryAction.js';
+import {
+  DataSpacePreviewAction,
+  DataSpacePreviewDialog,
+} from './DataSpacePreviewAction.js';
 
 const DATA_SPACE_ELEMENT_TYPE = 'DATA SPACE';
 const DATA_SPACE_ELEMENT_PROJECT_EXPLORER_DND_TYPE =
@@ -90,7 +92,7 @@ export class DSL_DataSpace_LegendStudioApplicationPlugin
       {
         key: 'promote-as-curated-template-query-page',
         addressPatterns: [
-          DATA_SPACE_QUERY_ROUTE_PATTERN.PROMOTE_TEMPLATE_QUERY,
+          DATA_SPACE_STUDIO_ROUTE_PATTERN.PROMOTE_TEMPLATE_QUERY,
         ],
         renderer: DataSpaceTemplateQueryPromotionReviewer,
       },
