+++ packages/legend-extension-dsl-data-space-studio/style/_data-space-studio-plugin.scss
