+++ packages/legend-extension-dsl-data-space-studio/style/_data-space-template-query-promotor.scss
