+++ packages/legend-extension-dsl-data-space-studio/tsconfig.build.json
@@ -7,16 +7,13 @@
   "exclude": ["src/**/__tests__/**/*.*", "src/**/__mocks__/**/*.*"],
   "references": [
     { "path": "./tsconfig.package.json" },
-    { "path": "../legend-graph/tsconfig.build.json" },
-    { "path": "../legend-extension-dsl-diagram/tsconfig.build.json" },
-    { "path": "../legend-application-query/tsconfig.build.json" },
-    { "path": "../legend-query-builder/tsconfig.build.json" },
+    { "path": "../legend-application/tsconfig.build.json" },
+    { "path": "../legend-application-studio/tsconfig.build.json" },
     { "path": "../legend-server-depot/tsconfig.build.json" },
     { "path": "../legend-server-sdlc/tsconfig.build.json" },
     { "path": "../legend-shared/tsconfig.build.json" },
-    { "path": "../legend-storage/tsconfig.build.json" },
+    { "path": "../legend-extension-dsl-data-space/tsconfig.build.json" },
     { "path": "../legend-application/tsconfig.build.json" },
-    { "path": "../legend-art/tsconfig.build.json" },
-    { "path": "../legend-lego/tsconfig.build.json" }
+    { "path": "../legend-art/tsconfig.build.json" }
   ]
 }
