+++ packages/legend-extension-dsl-data-space-studio/tsconfig.package.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@finos/legend-dev-utils/tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "lib",
+    "tsBuildInfoFile": "build/package.tsbuildinfo",
+    "rootDir": "."
+  },
+  "include": ["package.json"]
+}
