+++ packages/legend-extension-dsl-data-space-studio/tsconfig.publish.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.build.json",
+  "references": [{ "path": "./tsconfig.package.json" }]
+}
