+++ packages/legend-extension-dsl-data-space/src/components/shared/DSL_DataSpace_Icon.tsx
