+++ packages/legend-extension-dsl-data-space/src/stores/query-builder/DataSpaceQueryBuilderState.ts
@@ -51,9 +51,12 @@ import {
   type DataSpaceExecutionContext,
 } from '../../graph/metamodel/pure/model/packageableElements/dataSpace/DSL_DataSpace_DataSpace.js';
 import { DATA_SPACE_ELEMENT_CLASSIFIER_PATH } from '../../graph-manager/protocol/pure/DSL_DataSpace_PureProtocolProcessorPlugin.js';
-import { type DataSpaceInfo, extractDataSpaceInfo } from './DataSpaceInfo.js';
-import { DataSpaceAdvancedSearchState } from './DataSpaceAdvancedSearchState.js';
+import { DataSpaceAdvancedSearchState } from '../query/DataSpaceAdvancedSearchState.js';
 import type { DataSpaceAnalysisResult } from '../../graph-manager/action/analytics/DataSpaceAnalysis.js';
+import {
+  extractDataSpaceInfo,
+  type DataSpaceInfo,
+} from '../shared/DataSpaceInfo.js';
 
 export const resolveUsableDataSpaceClasses = (
   dataSpace: DataSpace,
