+++ packages/legend-extension-dsl-data-space/src/stores/query/DataSpaceQuerySetupStore.ts
@@ -41,7 +41,10 @@ import {
   UnsupportedOperationError,
   type GeneratorFn,
 } from '@finos/legend-shared';
-import { type DataSpaceInfo, extractDataSpaceInfo } from './DataSpaceInfo.js';
+import {
+  type DataSpaceInfo,
+  extractDataSpaceInfo,
+} from '../shared/DataSpaceInfo.js';
 import { DATA_SPACE_ELEMENT_CLASSIFIER_PATH } from '../../graph-manager/protocol/pure/DSL_DataSpace_PureProtocolProcessorPlugin.js';
 import { generateDataSpaceQueryCreatorRoute } from '../../__lib__/query/DSL_DataSpace_LegendQueryNavigation.js';
 import { renderDataSpaceQuerySetupSetupPanelContent } from '../../components/query/DataSpaceQuerySetup.js';
