+++ tsconfig.build.json
@@ -37,6 +37,7 @@
     { "path": "packages/legend-extension-dsl-text/tsconfig.build.json" },
     { "path": "packages/legend-extension-dsl-diagram/tsconfig.build.json" },
     { "path": "packages/legend-extension-dsl-data-space/tsconfig.build.json" },
+    { "path": "packages/legend-extension-dsl-data-space-studio/tsconfig.build.json" },
     { "path": "packages/legend-extension-dsl-service/tsconfig.build.json" },
     {
       "path": "packages/legend-extension-dsl-persistence/tsconfig.build.json"
