+++ tsconfig.json
@@ -49,6 +49,7 @@
     { "path": "packages/legend-extension-dsl-text" },
     { "path": "packages/legend-extension-dsl-diagram" },
     { "path": "packages/legend-extension-dsl-data-space" },
+    { "path": "packages/legend-extension-dsl-data-space-studio" },
     { "path": "packages/legend-extension-dsl-service" },
     { "path": "packages/legend-extension-dsl-persistence" },
     { "path": "packages/legend-extension-dsl-mastery" },
